Code improvements in model.py, dependency.py, and user.py, and add auth_utlis.py file that handles authentication-related logic
