+++ app/models.py
@@ -2,20 +2,11 @@ import re
 from typing import Optional
 from uuid import UUID, uuid4
 from datetime import datetime, timedelta, timezone
-from jose import JWTError, jwt
 from pydantic import BaseModel, EmailStr, Field, field_validator
 from passlib.context import CryptContext
-
-#Set up password hashing context using bcrypt
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-#JWT settings
-SECRET_KEY = "your_secret_key"
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 30
-
-password_regex = r"^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&#])[A-Za-z\d@$!%*?&#]{8,}$"
-
+import os
+from dotenv import load_dotenv
+load_dotenv()
 
 
 class UserBaseModel(BaseModel):
@@ -28,11 +19,14 @@ class UserCreateRequestModel(UserBaseModel):
     password: str
     @field_validator('password')
     def validate_password(cls, value):
-        if not re.match(password_regex, value):
+        if not re.match(os.getenv("password_regex"), value):
             raise ValueError(
                 "Password must be at least 8 characters long, contain at least 1 uppercase letter, 1 lowercase letter, 1 digit, and 1 special character."
             )
         return value
+    
+    def get_password_hash(self):
+        return os.getenv("pwd_context").hash(self.password)
 
 #Inheriting from UserBaseModel
 class User(UserBaseModel):
@@ -40,12 +34,10 @@ class User(UserBaseModel):
     created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
     updated_at: datetime = Field(default=None)
 
-    def get_password_hash(self):
-        return pwd_context.hash(self.password)
     
     #Method to verify if the provided password matches the hashed password
     def verify_password(self, password: str):
-        return pwd_context.verify(password, self.get_password_hash())
+        return os.getenv("pwd_context").verify(password, self.get_password_hash())
     
 
 
@@ -58,25 +50,3 @@ class TokenData(BaseModel):
     username: Optional[UUID] = None
 
 
-
-#JWT token creation
-def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.now(timezone.utc) + expires_delta  
-    else:
-        expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)  
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-#JWT token verification
-def verify_token(token: str, credentials_exception):
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        return username
-    except JWTError:
-        raise credentials_exception
